requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
EasyProcess==1.1
entrypoint2==1.1
evdev==1.7.0
idna==3.7
jeepney==0.8.0
markdown-it-py==3.0.0
mdurl==0.1.2
mss==9.0.1
pillow==10.3.0
Pygments==2.17.2
pynput==1.7.6
pyperclip==1.8.2
pyscreenshot==3.1
python-xlib==0.33
requests==2.31.0
rich==13.7.1
screeninfo==0.8.1
six==1.16.0
urllib3==2.2.1
